server/requirements.txt
alembic==1.14.0
asgiref==3.8.1
async-timeout==5.0.1
bcrypt==4.2.1
blinker==1.9.0
bpyutils==0.5.8
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Deprecated==1.2.15
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
flask-ldap3-login==1.0.2
Flask-Limiter==3.9.2
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
ldap3==2.9.1
limits==3.14.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==9.1.0
numpy==2.0.2
ordered-set==4.1.0
packaging==24.2
password-strength==0.0.3.post2
pillow==11.0.0
pipupgrade==1.12.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
qemu.qmp==0.0.3
qrcode==8.0
redis==5.2.1
requests==2.32.3
rich==13.9.4
setuptools==75.6.0
simplejson==3.19.3
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
websockify==0.12.0
Werkzeug==3.1.3
wrapt==1.17.0
WTForms==3.2.1
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
